- Handle JDBC as a resultset - Developed minimal app to test.
