+++ .gitignore
@@ -1,0 +1,14 @@
+/development/arx-poc/test_files_app/app_resultset/generate-resultset/nbproject/private/
+/development/arx-poc/test_files_app/app_resultset/generate-resultset/dist/
+/development/arx-poc/test_files_app/app_resultset/generate-resultset/build/
+/development/arx-poc/test_files_app/app_resultset/generate_resultset/dist/
+/development/arx-poc/test_files_app/app_resultset/resultset/build/
+/development/arx-poc/test_files_app/app_resultset/resultset/dist/
+/development/arx-poc/test_files_app/app_resultset/resultset/nbproject/private/
+/development/arx-poc/test_files_app/app_resultset/JavaLibrary1/nbproject/private/
+/development/PRIVAaaS-client/PRIVAaaS_Client/nbproject/private/
+/development/PRIVAaaS-client/PRIVAaaS_Client/dist/
+/development/PRIVAaaS-client/PRIVAaaS_Client/build/
+/development/PRIVAaaS/build/
+/development/PRIVAaaS/dist/
+/development/PRIVAaaS/nbproject/private/
