+++ development/PRIVAaaS-client/PRIVAaaS_Client/manifest.mf
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+X-COMMENT: Main-Class will be added automatically by build
+
