+++ development/PRIVAaaS-client/PRIVAaaS_Client/test_files/policy_ADULTO.csv
@@ -1,0 +1,16 @@
+id;1
+age;SR
+workclass;1
+fnlwgt;1
+education;1
+education2;1
+marital-status;SR
+occupation;1
+relationship;1
+race;1
+sex;1
+capital-gain;1
+capital-loss;1
+hours-per-week;1
+native-country;1
+Class;1
