--- development/PRIVAaaS-client/examples/README.md
@@ -1,8 +1,0 @@
-# examples
-
-
-usage examples:
-* curl -v http://localhost:4567/priva/ \
-     -d @"../../priva-poc/input/example-mock-data/mock_data.policy.json" \
-     -d @"../../priva-poc/input/example-mock-data/mock_data.json" \
-     -d @"../../priva-poc/input/example-mock-data/mock_data.json"
