--- development/PRIVAaaS-client/examples/demo_curl-url-enc_mock.sh
@@ -1,4 +1,0 @@
-curl -X POST -v http://localhost:4567/endpoint/ \
-     -d @"../../priva-poc/input/example-mock-data/mock_data.policy.json" \
-     -d @"../../priva-poc/input/example-mock-data/mock_data.json" \
-     -d @"../../priva-poc/input/example-mock-data/mock_data.json"
