+++ development/PRIVAaaS/manifest.mf
