+++ development/PRIVAaaS/test_files/param.csv
@@ -1,0 +1,6 @@
+ZIPCODE;SR
+DATETIME;1
+MIN;1
+MAX;SR
+COUNT;SR
+SUM;1
