--- development/arx-poc/nbproject/private/private.properties
@@ -1,7 +1,0 @@
-compile.on.save=false
-do.depend=false
-do.jar=true
-file.reference.libarx-3.5.1.jar-1=/home/marcio/projetos/atmosphere_repository/PRIVaaS/development/arx-poc/dist/lib/libarx-3.5.1.jar
-javac.debug=true
-javadoc.preview=true
-user.properties.file=/home/marcio/.netbeans/8.1/build.properties
