--- development/arx-poc/test_files/param.csv
@@ -1,6 +1,0 @@
-ZIPCODE;SR
-DATETIME;1
-MIN;1
-MAX;1
-COUNT;1
-SUM;1
